preserve the use of the "*"
Previous commits replaced the "*" with a 0. Code modified to not include
numerical zero vote counts where they have been redacted for privacy.
(Could be anywhere between 0-10 votes per precinct).

Per the Secretary of State elecion note: Note: In cases where the
cumulative turnout for a precinct for a race or ballot question is
greater than 0 but less than 10, the numbers have been replaced with an
asterisk (*) in order to protect the secrecy of a voter's ballot, as
required by Nevada Law. As a result, the total for a precinct may be
different from what is reported on official documents.